mage/requirements.txt
pyarrow
pandas
requests
dbt-core
dbt-bigquery
SQLAlchemy
fastparquet
google-cloud-bigquery
